Added redirects if user is not logged in.
- Added redirects if user is not logged in.
- Removed ";" from description in game view

+++ Backlog Checker/Controllers/GamesController.cs
@@ -47,32 +47,45 @@ namespace Backlog_Checker.Controllers
         [HttpGet]
         public IActionResult Game(int? gameId)
         {
-            if (gameId != null)
+            int? userId = HttpContext.Session.GetInt32("userId");
+
+            if (userId != null) 
             {
-                if (gamesManager.IfGameExists((int)gameId))
+                if (gameId != null)
                 {
-                    Game game = gamesManager.GetSingleGame((int)gameId);
-                    return View(game);
+                    if (gamesManager.IfGameExists((int)gameId))
+                    {
+                        Game game = gamesManager.GetSingleGame((int)gameId);
+                        return View(game);
+                    }
+                    else
+                    {
+                        return RedirectToAction("Index");
+                    }
+
                 }
                 else
                 {
                     return RedirectToAction("Index");
                 }
-                
             }
             else
             {
-                return RedirectToAction("Index");
-            }
-                
+                return RedirectToAction("Login", "Account");
+            }                
         }
 
         public void DeleteGame(int gameId)
         {
-            string rights = HttpContext.Session.GetString("rights");
             int? userId = HttpContext.Session.GetInt32("userId");
 
-            gamesManager.DeleteGame(rights, Convert.ToInt32(userId), gameId);
+            if (userId != null)
+            {
+                string rights = HttpContext.Session.GetString("rights");
+
+                gamesManager.DeleteGame(rights, Convert.ToInt32(userId), gameId);
+            }
+               
         }
 
         [HttpGet]
@@ -90,48 +103,65 @@ namespace Backlog_Checker.Controllers
         [HttpPost]
         public IActionResult AddGame(MyGamesViewModel model)
         {
-            GamesModelDTO DTO = new GamesModelDTO()
+            int? userId = HttpContext.Session.GetInt32("userId");
+
+            if (userId != null)
             {
-                Title = model.title,
-                Description = model.Description,
-                HeaderUrl = model.HeaderUrl,
-            };
-            try
-            {
-                gamesManager.AddGame(DTO);
-                return RedirectToAction("Index", "Games");
+                GamesModelDTO DTO = new GamesModelDTO()
+                {
+                    Title = model.title,
+                    Description = model.Description,
+                    HeaderUrl = model.HeaderUrl,
+                };
+                try
+                {
+                    gamesManager.AddGame(DTO);
+                    return RedirectToAction("Index", "Games");
+                }
+                catch
+                {
+                    throw new Exception("Failed to add game");
+                }
             }
-            catch
+            else
             {
-                throw new Exception("Failed to add game");
+                return RedirectToAction("Login", "Account");
             }
+
         }
 
         [HttpGet]
         public IActionResult Edit(int? gameId)
         {
             Game game;
+            int? userId = HttpContext.Session.GetInt32("userId");
 
-            if(gameId != null)
+            if (userId != null)
             {
-                string rights = HttpContext.Session.GetString("rights");
-                int? userId = HttpContext.Session.GetInt32("userId");
-
-                if (rights != "admin")
+                if (gameId != null)
                 {
+                    string rights = HttpContext.Session.GetString("rights");
 
-                    return RedirectToAction("Game", new { gameId = gameId });
+
+                    if (rights != "admin")
+                    {
+
+                        return RedirectToAction("Game", new { gameId = gameId });
+                    }
+
+                    game = gamesManager.GetSingleGame((int)gameId);
+
+                    return View(game);
                 }
-
-                game = gamesManager.GetSingleGame((int)gameId);
-
-                return View(game);
+                else
+                {
+                    return RedirectToAction("Index");
+                }
             }
             else
             {
-                return RedirectToAction("Index");
-            }
-            
+                return RedirectToAction("Login", "Account");
+            }           
         }
 
         [HttpPost]
@@ -139,15 +169,22 @@ namespace Backlog_Checker.Controllers
         {
             int? userId = HttpContext.Session.GetInt32("userId");
 
-            game.UpdateGame();
+            if (userId != null)
+            {
+                game.UpdateGame();
 
-            return RedirectToAction("Game", new { gameId = game.Id });
+                return RedirectToAction("Game", new { gameId = game.Id });
+            }
+            else
+            {
+                return RedirectToAction("Login", "Account");
+            }
         }
 
         public void ToggleOwned(int gameId, string subject)
         {
+
             int? userId = HttpContext.Session.GetInt32("userId");
-            
 
             gamesManager.ToggleUserGameRelation(gameId, subject, Convert.ToInt32(userId));
         }
